Refactor compactor's profiler record naming and adjust chunk factor
Refactored the code for better readability by moving the creation of the name used in profiler records to the NaiveCompactor constructor. Also replaced the commented-out statement with an if-statement in compactor.cpp for chunk count checking. In main.cpp, the chunk factor was reduced from 16 to 1, which optimizes the size of processing chunks to improve program performance.