Update Profiling flags and Add Reset method to DataChunk
The kEnableProfiling flag was changed from true to false for BeeProfiler and from false to true for ZebraProfiler inside profiler.h. A Reset method was added to the DataChunk class in base.h that resets its count and reinitializes its selection_vector. This will allow for efficient object reuse. The "swap" operation method in compactor.cpp was also rewritten for better clarity and performance.