Add BinaryCompactor class and update usage of Profiler
Introduced a new BinaryCompactor class and made modifications to existing classes and scripts. The Profiler instance is now a member of compactor classes, instead of being locally declared in functions. Also changed the type of compactor in the main script to newly introduced BinaryCompactor and increased the 'kChunkFactor' from 4 to 9 consequently.