Add payload_length parameter to HashTable constructor
The HashTable constructor now accepts an additional `payload_length` parameter. This accommodates for different payload lengths in each join operation. Setting files have been updated to reflect these changes and the use of different compactor types. The handling of compactor types is now also under the compaction setting in the setting.h file, instead of in the main.cpp file.