Change executable name and add build script
This commit primarily changes the name of the generated executable from `exe_logical_compact` to `compaction`. In addition, a new shell script `build_versions.sh` is added, which automates the building process of the `compaction` project. This script also moves the resulting compiled project into a specific directory for better organization.